Routing setup and Navigation from login screen to home screen
