+++ lib/constants.dart
@@ -1,0 +1,2 @@
+const String loginRoute = '/';
+const String homeRoute = '/home';
