+++ pubspec.yaml
@@ -46,7 +46,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
      - assets/images/bg_image.jpg
-
+     
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
